Actions

+++ src/app/actions/auth.ts
@@ -6,7 +6,7 @@ import User from '@/src/models/User'
 import argon2id from 'argon2'
 import { cookies, headers } from 'next/headers'
 import { redirect } from 'next/navigation'
-import { UserTypes } from '../types/User'
+import { UserType } from '../types/User'
 import env from '@/src/lib/env'
 
 async function login({ email, password }: { email: string, password: string }) {
@@ -51,25 +51,25 @@ async function logout() {
 	return { message: 'Successfully logged out' }
 }
 
-async function validateSession(allowedRoles?: UserTypes['role'][]) {
-	if (!cookies().get('session')) redirect('/')
+async function validateSession(allowedRoles?: UserType['role'][], redirectUser?: boolean) {
+	if (!cookies().get('session')) redirectUser ? redirect('/') : { loggedIn: false, role: 'guest' }
 
 	dbConnect();
 	const session = await Session.findById(cookies().get('session')?.value)
 
 	// If the session doesn't exist or the user agent doesn't match, delete the session and redirect to the login page
 	if (!session || headers().get('user-agent') !== session.userAgent) {
-		await session.delete()
+		if (session) await session.delete()
 		cookies().set('session', '', { expires: new Date(0), httpOnly: true, path: '/', sameSite: 'strict', secure: process.env.NODE_ENV === 'production' })
-		return redirect('/')
+		return redirectUser ? redirect('/') : { loggedIn: false, role: 'guest' }
 	}
 
+	const user = await User.findById(session.userID)
 	if (allowedRoles) {
-		const user = await User.findById(session.userID)
-		if (!allowedRoles.includes(user.role)) return redirect('/')
+		if (!allowedRoles.includes(user.role)) return redirectUser ? redirect('/') : { loggedIn: false, role: 'guest' }
 	}
 
-	return
+	return { loggedIn: true, role: user.role }
 }
 
 export { login, logout, validateSession }
